SRS: Updated README with proper table formatting
Previously, the tables example at the bottom of the README used the
tabular environment, but the SRS document should instead use the
tabularx environment, as it allows tables to easily fill the width of
the page.

This can be accomplished by using the `X` option (as opposed to r, l, or
c align) when specifying a column. For example,
  \begin{tabularx}{\textwidth}{lX}
creates two columns. The left column is text-aligned to the left and the
right column spans the leftover width on the page.