WEB: Initial prototype of file viewer
Completely reworked the front-end so that queries result in a
file-system tree layout within the QueryResults component. Allows for
context, i.e. the ability to search within a specific part of the tree.
Back-button functionality is included, so uses can use the back button
to navigate up the file tree. In the future, should include the ability
to make jumps.

Added a basic FileViewer component, which should be fleshed out.